Made several more or less usefull changes, mostly optimizations.
In BaseThing: saw to that GetNote uses the more up to date module random,
instead of whrandom. Also changed it to use random.choice. We use a list now
instead of assigning integers to the choices.

GetListLine: added  a dash - to be prepended to a listline, not just a name.
thought it would look better if all lines where the same in invetory and such.

Removed the PUB-version off replace and replaced all occurences of it
with string.replace in some manner.

Added some checks in restoregave and savegame, to see if files exist, if games
should be saved and some such.