Added Eat and Drink
  pubobjs:
Added Edible and Drinkable
Added a few line in Room.GetDesc to enable the use of initialDesc on rooms.
Added hunger and thirst to Actor()

  pubverbs:
Added the verbs Eat and Drink as well as Script.
Script can be deleted as soon as pubdebug is inserted, I just need it for
debugging in the meantime.

  pubcore:
Added som code to BaseThing.MoveTo to enable the use of 'TRASH'
used to delete objects.

Eat and Drink are not all they should or could be yet.
If you drink a sea you could empy it in one go.
I'll fix this later.