Add support for unit testing.
This test harness looks in the 'test' directory for any .py files that
start with the string 'test'.  It assumes that such a file is a module
containing a member 'suite' that represents a TestSuite.  It imports
that suite, adds it to a master TestSuite, and then uses unittest's
TextTestRunner to run the all of the tests in a single run.  So to
add a new set of unit tests, just make a new testXXXX.py file, and
make sure its 'suite' variable is set to a TestSuite that references
all the TestCases mentioned in the file.  See testDebug.py as an
example.