Major name changes:
What was called "Symbol" was a container for
dynamic components. It is now called DynamicComponentRack, and
the interface which was ISymbol is now IDynamicComponentRack.

What was called "Concept" is now "Symbol". At this time there is
no ISymbol. Symbols don't have much of an interface anyway -- the
whole point is that they are essentially bare names with no
associations. They are used like enumerations or numbers -- as
atomic objects. What limited interface they provide is for the
convenience of testing, debugging, documentation, and translation.
They are maintained through a central registry: Symbol.sym
which other modules import.

Intermediate progress on defining semantics and the core
localization.

Writing and testing English locale (locales/en.py) grammar, and
needed utilities in locales/l10n.py and locales/tests.py
Haven't started on the XML vocabulary reader, though I've decided
to use elementtree to process it.

Some cleanup of older modules.